components/tools/esp_prov/requirements.txt
future
protobuf
netifaces
dbus-python
cryptography
